add validation config
